refactor: type을 interface로 변경
