+++ features/soft-skills/api/fetch-spectrum-by-project.api.ts
@@ -1,6 +1,6 @@
 import { apiRequest } from '@/shared/lib/api';
 
-import type { SpectrumInfo } from '../types/soft-skills.types';
+import type { SpectrumInfo } from '../soft-skills.types';
 
 export const fetchSpectrumByProject = (userId: number, projectId: number): Promise<SpectrumInfo> =>
   apiRequest<SpectrumInfo>(`/reviews/spectrums/${userId}/${projectId}`);
