+++ features/soft-skills/components/soft-skills-spectrum.tsx
@@ -1,4 +1,4 @@
-import type { SpectrumInfo } from '../types/soft-skills.types';
+import type { SpectrumInfo } from '../soft-skills.types';
 
 import { spectrumAxisLabels } from '../constants/spectrum-axis-labels';
 import { toMarkerLeft } from '../utils/to-marker-left';
@@ -12,7 +12,11 @@ const SPECTRUM_TRACK_HEIGHT_PX = 136;
 /** 트랙 위 세로 점선 가이드 위치(%) — 5구간 시각 보조 */
 const SPECTRUM_VERTICAL_DASH_PERCENTS = [20, 40, 60, 80] as const;
 
-const SpectrumVerticalDashOverlay = ({ widthPx }: { widthPx: number }) => (
+interface SpectrumVerticalDashOverlayProps {
+  widthPx: number;
+}
+
+const SpectrumVerticalDashOverlay = ({ widthPx }: SpectrumVerticalDashOverlayProps) => (
   <svg
     aria-hidden
     className="pointer-events-none absolute top-0 left-0 select-none"
@@ -29,7 +33,11 @@ const SpectrumVerticalDashOverlay = ({ widthPx }: { widthPx: number }) => (
   </svg>
 );
 
-export const SoftSkillsSpectrum = ({ spectrumInfo }: { spectrumInfo: SpectrumInfo }) => (
+interface SoftSkillsSpectrumProps {
+  spectrumInfo: SpectrumInfo;
+}
+
+export const SoftSkillsSpectrum = ({ spectrumInfo }: SoftSkillsSpectrumProps) => (
   <div className="w-fill mt-7">
     <h3 className="text-heading-xs text-text-subtle h-6 font-bold">평균 지표</h3>
