+++ features/soft-skills/hooks/use-spectrum.ts
@@ -4,12 +4,12 @@ import { fetchSpectrumByProject } from '../api/fetch-spectrum-by-project.api';
 import { fetchSpectrum } from '../api/fetch-spectrum.api';
 import { softSkillsKeys } from '../soft-skills.query-key';
 
-interface Params {
+interface UseSpectrumParams {
   userId: number;
   projectId?: number;
 }
 
-export const useSpectrum = ({ userId, projectId }: Params) =>
+export const useSpectrum = ({ userId, projectId }: UseSpectrumParams) =>
   useSuspenseQuery({
     queryKey: softSkillsKeys.spectrum(userId, projectId),
     queryFn: () => (projectId === undefined ? fetchSpectrum(userId) : fetchSpectrumByProject(userId, projectId)),
